HBX-1563: Move class 'org.hibernate.cfg.reveng.JDBCToHibernateTypeHelper' to 'org.hibernate.tool.internal.util.JdbcToHibernateTypeHelper'
