+++ main/src/main/java/org/hibernate/cfg/JDBCBinder.java
@@ -28,7 +28,6 @@ import org.hibernate.cfg.binder.PrimaryKeyInfo;
 import org.hibernate.cfg.binder.PropertyBinder;
 import org.hibernate.cfg.reveng.DatabaseCollector;
 import org.hibernate.cfg.reveng.JDBCReader;
-import org.hibernate.cfg.reveng.JDBCToHibernateTypeHelper;
 import org.hibernate.cfg.reveng.MappingsDatabaseCollector;
 import org.hibernate.cfg.reveng.RevEngUtils;
 import org.hibernate.cfg.reveng.ReverseEngineeringStrategy;
@@ -55,6 +54,7 @@ import org.hibernate.mapping.Table;
 import org.hibernate.mapping.ToOne;
 import org.hibernate.service.ServiceRegistry;
 import org.hibernate.tool.api.reveng.AssociationInfo;
+import org.hibernate.tool.internal.util.JdbcToHibernateTypeHelper;
 import org.hibernate.tool.util.TableNameQualifier;
 import org.hibernate.type.ForeignKeyDirection;
 import org.hibernate.type.Type;
@@ -925,7 +925,7 @@ public class JDBCBinder {
 	}
 
 	private String typeCodeName(int sqlTypeCode) {
-		return sqlTypeCode + "(" + JDBCToHibernateTypeHelper.getJDBCTypeName(sqlTypeCode) + ")";
+		return sqlTypeCode + "(" + JdbcToHibernateTypeHelper.getJDBCTypeName(sqlTypeCode) + ")";
 	}
 
 	/**
