+++ main/src/main/java/org/hibernate/cfg/reveng/BasicColumnProcessor.java
@@ -9,6 +9,7 @@ import org.hibernate.cfg.JDBCBinderException;
 import org.hibernate.cfg.reveng.dialect.MetaDataDialect;
 import org.hibernate.mapping.Column;
 import org.hibernate.mapping.Table;
+import org.hibernate.tool.internal.util.JdbcToHibernateTypeHelper;
 import org.hibernate.tool.util.TableNameQualifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -83,15 +84,15 @@ public class BasicColumnProcessor {
 				column.setComment(comment);
 				column.setSqlTypeCode(new Integer(sqlType) );
                 if(intBounds(size) ) {
-                	if(JDBCToHibernateTypeHelper.typeHasLength(sqlType) ) {
+                	if(JdbcToHibernateTypeHelper.typeHasLength(sqlType) ) {
                 		column.setLength(size);
                 	} 
-                	if(JDBCToHibernateTypeHelper.typeHasScaleAndPrecision(sqlType) ) {
+                	if(JdbcToHibernateTypeHelper.typeHasScaleAndPrecision(sqlType) ) {
                 		column.setPrecision(size); 
                 	}
 				} 
                 if(intBounds(decimalDigits) ) {
-                	if(JDBCToHibernateTypeHelper.typeHasScaleAndPrecision(sqlType) ) {
+                	if(JdbcToHibernateTypeHelper.typeHasScaleAndPrecision(sqlType) ) {
                 		column.setScale(decimalDigits);
                 	}
 				}
